Migrating map diff function and structs to utils
