+++ iDiff/differs/aptDiff.go
@@ -2,40 +2,16 @@ package differs
 
 import (
 	"bufio"
+	"github.com/runtimes-common/iDiff/utils"
 	"html/template"
 	"log"
 	"os"
-	"reflect"
 	"strings"
-
-	"github.com/runtimes-common/iDiff/utils"
 
 	"github.com/golang/glog"
 )
 
-// PackageDiff stores the difference information between two images.
-type PackageDiff struct {
-	Image1    string
-	Packages1 []string
-	Image2    string
-	Packages2 []string
-	InfoDiff  []Info
-}
-
-// Info stores the information for one package in two different images.
-type Info struct {
-	Package string
-	Info1   PackageInfo
-	Info2   PackageInfo
-}
-
-// PackageInfo stores the specific metadata about a package.
-type PackageInfo struct {
-	Version string
-	Size    string
-}
-
-func output(diff PackageDiff) error {
+func output(diff utils.PackageDiff) error {
 	const master = `Packages found only in {{.Image1}}:{{block "list" .Packages1}}{{"\n"}}{{range .}}{{println "-" .}}{{end}}{{end}}
 Packages found only in {{.Image2}}:{{block "list2" .Packages2}}{{"\n"}}{{range .}}{{println "-" .}}{{end}}{{end}}
 Version differences:{{"\n"}}	(Package:	{{.Image1}}{{"\t\t"}}{{.Image2}}){{range .InfoDiff}}
@@ -79,41 +55,15 @@ func AptDiff(d1file, d2file string) (string, error) {
 		return "", err
 	}
 
-	diff := diffMaps(pack1, pack2)
+	diff := utils.DiffMaps(pack1, pack2)
 	diff.Image1 = dirPath1
 	diff.Image2 = dirPath2
 	output(diff)
 	return "", nil
 }
 
-func (pi PackageInfo) string() string {
-	return pi.Version
-}
-
-func diffMaps(map1, map2 map[string]PackageInfo) PackageDiff {
-	diff1 := []string{}
-	diff2 := []string{}
-	infoDiff := []Info{}
-	for key1, value1 := range map1 {
-		value2, ok := map2[key1]
-		if !ok {
-			diff1 = append(diff1, key1+":"+value1.string())
-		} else if !reflect.DeepEqual(value2, value1) {
-			infoDiff = append(infoDiff, Info{key1, value1, value2})
-			delete(map2, key1)
-		} else {
-			delete(map2, key1)
-		}
-	}
-	for key2, value2 := range map2 {
-		diff2 = append(diff2, key2+":"+value2.string())
-	}
-	diff := PackageDiff{Packages1: diff1, Packages2: diff2, InfoDiff: infoDiff}
-	return diff
-}
-
-func getPackages(path string) (map[string]PackageInfo, error) {
-	packages := make(map[string]PackageInfo)
+func getPackages(path string) (map[string]utils.PackageInfo, error) {
+	packages := make(map[string]utils.PackageInfo)
 	layerStems, err := utils.BuildLayerTargets(path, "layer/var/lib/dpkg/status")
 	if err != nil {
 		return packages, err
@@ -140,7 +90,7 @@ func getPackages(path string) (map[string]PackageInfo, error) {
 	return packages, nil
 }
 
-func parseLine(text string, currPackage string, packages map[string]PackageInfo) string {
+func parseLine(text string, currPackage string, packages map[string]utils.PackageInfo) string {
 	line := strings.Split(text, ": ")
 	if len(line) == 2 {
 		key := line[0]
@@ -155,12 +105,12 @@ func parseLine(text string, currPackage string, packages map[string]PackageInfo)
 				return currPackage
 			}
 			modifiedValue := strings.Replace(value, "+", " ", 1)
-			tempPackage := PackageInfo{Version: modifiedValue}
+			tempPackage := utils.PackageInfo{Version: modifiedValue}
 			packages[currPackage] = tempPackage
 			return currPackage
 
 		case "Installed-Size":
-			tempPackage := PackageInfo{Size: value}
+			tempPackage := utils.PackageInfo{Size: value}
 			packages[currPackage] = tempPackage
 			return currPackage
 		default:
