+++ iDiff/utils/package_diff_utils_test.go
@@ -6,6 +6,87 @@ import (
 	"testing"
 )
 
+type ByPackage []Info
+
+func (a ByPackage) Len() int {
+	return len(a)
+}
+
+func (a ByPackage) Swap(i, j int) {
+	a[i], a[j] = a[j], a[i]
+}
+
+func (a ByPackage) Less(i, j int) bool {
+	return a[i].Package < a[j].Package
+}
+
+func TestDiffMaps(t *testing.T) {
+	testCases := []struct {
+		descrip  string
+		map1     map[string]PackageInfo
+		map2     map[string]PackageInfo
+		expected PackageDiff
+	}{
+		{
+			descrip: "Missing Packages.",
+			map1: map[string]PackageInfo{
+				"pac1": PackageInfo{"1.0", "40"},
+				"pac3": PackageInfo{"3.0", "60"}},
+			map2: map[string]PackageInfo{
+				"pac4": PackageInfo{"4.0", "70"},
+				"pac5": PackageInfo{"5.0", "80"}},
+			expected: PackageDiff{
+				Packages1: []string{"pac1:1.0", "pac3:3.0"},
+				Packages2: []string{"pac4:4.0", "pac5:5.0"},
+				InfoDiff:  []Info{},
+			},
+		},
+		{
+			descrip: "Different Versions and Sizes.",
+			map1: map[string]PackageInfo{
+				"pac2": PackageInfo{"2.0", "50"},
+				"pac3": PackageInfo{"3.0", "60"}},
+			map2: map[string]PackageInfo{
+				"pac2": PackageInfo{"2.0", "45"},
+				"pac3": PackageInfo{"4.0", "60"}},
+			expected: PackageDiff{
+				Packages1: []string{},
+				Packages2: []string{},
+				InfoDiff: []Info{Info{"pac2", PackageInfo{"2.0", "50"}, PackageInfo{"2.0", "45"}},
+					Info{"pac3", PackageInfo{"3.0", "60"}, PackageInfo{"4.0", "60"}}},
+			},
+		},
+		{
+			descrip: "Identical packages, versions, and sizes",
+			map1: map[string]PackageInfo{
+				"pac1": PackageInfo{"1.0", "40"},
+				"pac2": PackageInfo{"2.0", "50"},
+				"pac3": PackageInfo{"3.0", "60"}},
+			map2: map[string]PackageInfo{
+				"pac1": PackageInfo{"1.0", "40"},
+				"pac2": PackageInfo{"2.0", "50"},
+				"pac3": PackageInfo{"3.0", "60"}},
+			expected: PackageDiff{
+				Packages1: []string{},
+				Packages2: []string{},
+				InfoDiff:  []Info{},
+			},
+		},
+	}
+	for _, test := range testCases {
+		diff := DiffMaps(test.map1, test.map2)
+		sort.Strings(test.expected.Packages1)
+		sort.Strings(diff.Packages1)
+		sort.Strings(test.expected.Packages2)
+		sort.Strings(diff.Packages2)
+		sort.Sort(ByPackage(test.expected.InfoDiff))
+		sort.Sort(ByPackage(diff.InfoDiff))
+		if !reflect.DeepEqual(test.expected, diff) {
+			t.Errorf("Expected packages only in map1 to be: %s but got: %s", test.expected, diff)
+		}
+	}
+}
+
 func TestBuildLayerTargets(t *testing.T) {
 	path := "test_files/dir1"
 	target := "123"
